--- a/app/views/checklist_tests/_checklist_measure.html.erb
+++ b/app/views/checklist_tests/_checklist_measure.html.erb
@@ -33,7 +33,7 @@
             <tbody>
               <%= f.fields_for :checked_criteria, product_test.checked_criteria.all(measure_id: measure.id) do |criteria_field| %>
                 <% criteria = measure.hqmf_document[:data_criteria].select { |key, value| key == criteria_field.object.source_data_criteria }.values.first %>
-                <% if criteria.has_key?('description') %>
+                <% if criteria.key?('description') %>
                   <tr>
                     <td rowspan="3">
                       <% if criteria_field.object.passed_qrda %>