--- a/app/views/vendors/index.html.erb
+++ b/app/views/vendors/index.html.erb
@@ -24,37 +24,7 @@
       </tr>
     </thead>
     <tbody>
-      <% @vendors.each do |vendor| %>
-        <% vendor_statuses = vendor_statuses(vendor) %>
-        <tr>
-          <th scope="row"><div class = "abbreviated"><%= link_to vendor.name, vendor_path(vendor) %></div></th>
-          <td>
-            <% if vendor_statuses['total'] == 0 %>
-              <%= button_to new_vendor_product_path(vendor), :method => :get, :class => "btn btn-primary btn-sm" do %>
-                <i class="fa fa-fw fa-plus" aria-hidden="true"></i> Add Product
-              <% end %>
-            <% else %>
-              <%= vendor_statuses['total'] %>
-            <% end %>
-          </td>
-
-          <% %w(passing failing errored incomplete).each do |status| %>
-            <% if vendor_statuses[status] > 0 %>
-              <% classes = status_to_css_classes(status) %>
-              <td class="<%= classes['cell'] %> text-center">
-                <%= vendor_statuses[status] %>
-              </td>
-            <% else %>
-              <td></td>
-            <% end %>
-          <% end %>
-          <td class="text-right">
-            <%= button_to edit_vendor_path(vendor), :method => :get, :class => "btn btn-sm btn-default" do %>
-              <i class="fa fa-fw fa-wrench" aria-hidden="true"></i> Edit Vendor
-            <% end %>
-          </td>
-        </tr>
-      <% end %>
+      <%= render partial: 'vendor', collection: @vendors %>
     </tbody>
   </table>
 <% end %>