--- a/app/views/vendors/show.html.erb
+++ b/app/views/vendors/show.html.erb
@@ -9,9 +9,19 @@
   </p>
 <% else %>
   <h1><%= @products.count %> Products</h1>
-  <div class="vendor-product-status">
-    <% @products.each do |product| %>
-      <%= render partial: 'product_status_table', locals: { show_product_link: true, product: product } %>
-    <% end %>
+  <div class="vendor-product-status infinite-scroll">
+    <%= render partial: 'product_status_table', collection: @products, as: :product, locals: { show_product_link: true } %>
   </div>
+  <% unless @products.current_page == @products.total_pages %>
+    <div class="text-center scroll-link-wrapper">
+      <p id="view-more">
+        <!-- Error should only show up when the page is not properly infinitely scrolling -->
+        <%= link_to('Loading Error - Try Again?', url_for(page: @products.current_page + 1)) %>
+      </p>
+      <p id="loading-more" class="default-hidden">
+        <i class="fa fa-fw fa-refresh fa-spin inline-icon info-icon"></i>
+        Loading More Products
+      </p>
+    </div>
+  <% end %>
 <% end %>