This PR has two different parts to it. The first part is some optimizations to make the vendor and product pages load faster. The result of the optimizations in the first commit are as follows:

> Without any optimizations: 112516ms (10340 sql queries in vendors)
> With "includes" optimizations 80057ms (6384 sql queries in vendors)

Additionally the tmp:clear rake task is now enhanced to also rebuild the cache right away after it is cleared. This means that in reality after running tmp:clear (which is run after every upgrade) the load time should be about `7868.5ms` with no sql queries run. I believe the additional time is due to the cache being actually loaded into memory. Subsequent reloads are in the `< 1s` range.

The second part of this is the infinite scrolling. Due to the design of the vendors page it was not possible to implement infinite scrolling on it, the problem being that sorting would no longer work correctly if additional elements were being added as the user scrolled. This problem was mitigated by the above optimizations. For the products page it was possible to implement actual infinite scrolling which helped the performance of that page on a cold cache considerably.

At this point I have not implemented any tests for the infinite scrolling code, if those are desired I can attempt to add those in as well.

@ehathaway The reason I included you on this specifically is due to the product infinite scrolling since it does affect the UI behavior a bit.