@rbclark 2016-12-15
Converted to js, everything seems to still work fine.